refactored asset management
