+++ ass_man/serializers/network_port_serializers.py
@@ -1,0 +1,8 @@
+from ass_man.models import Network_Port
+from rest_framework import serializers
+
+
+class NetworkPortSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Network_Port
+        fields = ['name', 'mac', 'connection']
