+++ ass_man/serializers/pdu_serializers.py
@@ -1,0 +1,8 @@
+from ass_man.models import PDU
+from rest_framework import serializers
+
+
+class PDUSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = PDU
+        fields = ['name']
