+++ backend/urls.py
@@ -16,22 +16,21 @@ from django.contrib import admin
 from django.urls import path, include, re_path
 from rest_framework import routers
 from django.views.generic import TemplateView
-import ass_man.views as asset_views
 from usr_man import views as user_views
-from ass_man.views import report
+from ass_man.views import ReportView, ModelViewSet, AssetViewSet, RackViewSet, DatacenterViewSet
 
 router = routers.DefaultRouter()  # add this
 router.register(r'users', user_views.UserViewSet)
-router.register(r'models', asset_views.ModelViewSet)
-router.register(r'assets', asset_views.AssetViewSet)
-router.register(r'racks', asset_views.RackViewSet)
-router.register(r'datacenters', asset_views.DatacenterViewSet)
+router.register(r'models', ModelViewSet.ModelViewSet)
+router.register(r'assets', AssetViewSet.AssetViewSet)
+router.register(r'racks', RackViewSet.RackViewSet)
+router.register(r'datacenters', DatacenterViewSet.DatacenterViewSet)
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include('django.contrib.auth.urls')),
     path('api/', include(router.urls)),  # add this
-    path('report/', report),
+    path('report/', ReportView.report),
     path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
     re_path('.*', TemplateView.as_view(template_name='index.html')),
 ]
